+++ src/contentScript/enhance.js
@@ -3,7 +3,7 @@ const enhanceMarkAttributeName = "data-translationmark";
 const enhanceOriginalDisplayValueAttributeName = "data-translationoriginaldisplay";
 const enhanceHtmlTagsInlineIgnore = ['BR', 'CODE', 'KBD', 'WBR'] // and input if type is submit or button, and pre depending on settings
 const enhanceHtmlTagsNoTranslate = ['TITLE', 'SCRIPT', 'STYLE', 'TEXTAREA', 'SVG', 'svg'] //TODO verificar porque 'svg' é com letras minúsculas
-const blockElements = [
+let blockElements = [
     'H1', 'H2', 'H3', 'H4', 'H5', 'H6','TABLE',  'OL', 'P','LI'
   ];
 if (twpConfig.get('translateTag_pre') !== 'yes') {
@@ -14,12 +14,8 @@ const headingElements = ['h1' ];
 
 const pdfSelectorsConfig =   {
     regex:
-      "translatewebpages.org/result/.+$",
-    selectors:[
-      'div'
-    ],
-    style:"none",
-  };
+      "translatewebpages.org/result/.+$"
+};
 
 const inlineElements = [
   "a",
@@ -57,160 +53,6 @@ const inlineElements = [
   "var",
 ];
 
-const translateSelectors = [
-  {
-    hostname:["twitter.com","tweetdeck.twitter.com","mobile.twitter.com"],
-    selectors:[
-     '[data-testid="tweetText"]',".tweet-text" ,".js-quoted-tweet-text"
-    ]
-  },
-  {
-    hostname:"news.ycombinator.com",
-    selectors:[
-      ".titleline >a",
-      '.comment',
-      '.toptext'
-
-    ]
-  },
-  {
-    hostname:"www.reddit.com",
-    selectors:[
-      // "[data-adclicklocation=title]",
-      "h1",
-      "[data-click-id=body] h3","[data-click-id=background] h3"
-    ],
-    containerSelectors:[
-      "[data-testid=comment]",
-      "[data-adclicklocation=media]"
-    ]
-  },
-  {
-    hostname:"old.reddit.com",
-    selectors:[
-      "a.title",
-    ],
-    containerSelectors:[
-      "[role=main] .md-container"
-    ]
-  },
-  {
-    regex:"finance\.yahoo\.com\/$",
-    selectors:[
-      "h3"
-    ]
-
-  },
-  {
-    regex:["www\.bloomberg\.com\/[A-Za-z0-9]+$","www\.bloomberg\.com\/$"],
-    selectors:[
-      "article h3",
-      "article .single-story-module__headline-link",
-      "article [data-tracking-type=Story]",
-      "article .story-list-story__info__headline"
-
-    ]
-  },
-  pdfSelectorsConfig,
-  {
-    hostname:"www.cell.com",
-    selectors:[
-      "div.section-paragraph > div.section-paragraph > div.section-paragraph",
-      "section > div.section-paragraph",
-      "h4","h3","h2"
-    ],
-  },
-  {
-    // TODO
-    hostname:["www.msdmanuals.com","localhost"],
-    noTranslateSelectors:[
-      ".d-none"
-    ]
-  },
-  {
-    hostname:"www.reuters.com",
-    containerSelectors:'main',
-  },
-  {
-    regex:"finance\.yahoo\.com/news",
-    containerSelectors:"[role=article]"
-  },
-  {
-    hostname:"www.whatsonweibo.com",
-    containerSelectors:"#mvp-post-main"
-  },
-  {
-    hostname:["www.wsj.com","www.economist.com"],
-    containerSelectors:"main"
-  },
-
-  {
-    hostname:["mail.jabber.org","antirez.com"],
-    selectors:["pre"],
-    containerSelectors: "pre",
-    style: 'none'
-  },
-  {
-    hostname:"github.com",
-    containerSelectors:".markdown-body"
-  },
-  {
-    hostname:"www.youtube.com",
-    selectors:["#content-text"]
-  },
-  {
-    hostname:"www.facebook.com",
-    selectors:["div[data-ad-comet-preview=message] > div > div","div[role=article] > div > div > div > div > div > div > div > div "]
-    
-  },
-  {
-    regex:'\.substack\.com\/',
-    selectors:[
-      ".post-preview-title",
-      ".post-preview-description"
-    ],
-    containerSelectors:[
-      ".post"
-    ]
-  },
-  {
-    hostname:"www.nature.com",
-    containerSelectors:"article"
-  },{
-    hostname:"seekingalpha.com",
-    containerSelectors:"div.wsb_section",
-    brToParagraph: true
-  },{
-    hostname:"hn.algolia.com",
-    selectors:[".Story_title"]
-  },{
-    hostname:"read.readwise.io",
-    selectors:['div[class^="_titleRow_"]','div[class^="_description_"]'],
-    containerSelectors:[
-      "#document-text-content"
-    ]
-  },{
-    hostname:"www.inoreader.com",
-    selectors:[".article_title",],
-    containerSelectors:[
-      ".article_content"
-    ]
-  },
-  {
-    hostname:"mail.google.com",
-    selectors:["h2[data-thread-perm-id]","span[data-thread-id]"],
-    containerSelectors:[
-      "div[data-message-id] > div > div[class='']"
-    ]
-  },{
-    hostname:"www.producthunt.com",
-    selectors:["a[data-test^='post-']",'h2',"div.layoutCompact div[class^='styles_htmlText__']"],
-  },{
-    hostname:"arxiv.org",
-    selectors:["blockquote.abstract","h1"]
-  }
-
-]
 
 function addWrapperToNode(node, wrapper){
   try{
@@ -231,9 +73,12 @@ function getPageSpecialConfig(ctx){
   const currentUrlObj = new URL(currentUrl);
   const currentHostname = currentUrlObj.hostname;
   const currentUrlWithoutSearch = currentUrlObj.origin + currentUrlObj.pathname;
+
+  // merge spcialRules
+
   let specialConfig = null;
 
-  for(const enhance of translateSelectors){
+  for(const enhance of specialRules){
     if(enhance.hostname){
       if(!Array.isArray(enhance.hostname)){
         enhance.hostname = [enhance.hostname];
@@ -365,6 +210,10 @@ async function getNodesThatNeedToTranslate(root,ctx,options){
   const currentHostname = currentUrlObj.hostname;
   let currentTargetLanguage = twpConfig.get("targetLanguage")
 
+  // special for mail.google.com, cause there are too many table, we should remove table
+  if(pageSpecialConfig && pageSpecialConfig.blockElements){
+    blockElements = pageSpecialConfig.blockElements;
+  }
   // check sites
   if(allBlocksSelectors.length>0){
     for(const selector of allBlocksSelectors){